@@ -8,7 +8,7 @@
     "authors": [
         {
             "name": "Nelson da Costa",
-            "email": "baruica@gmail.com"
+            "homepage": "https://twitter.com/baruica"
         }
     ],
     "autoload": {
@@ -26,16 +26,13 @@
         }
     },
     "config": {
-        "platform": {
-            "php": "7.0"
-        },
         "sort-packages": true
     },
     "require": {
-        "php": "^7.0"
+        "php": "^7.4",
+        "ext-dom": "*"
     },
     "require-dev": {
-        "phpspec/phpspec": "^4.0",
-        "phpstan/phpstan": "^0.8.0"
+        "phpspec/phpspec": "^6.1"
     }
 }
